refactor: outgoing, gcp, and twitter utilities
